requirements.txt
boto3==1.21.40
botocore==1.24.40
imageio==2.16.1
numpy==1.22.3
opencv_contrib_python==4.5.5.64
Shapely==1.8.2
imageio-ffmpeg~=0.4.5
requests~=2.27.1
pandas
Flask
Pillow~=9.1.0
opencv-python~=4.5.5.64
requests
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
torchaudio
astunparse
expecttest
future
numpy
psutil
pyyaml
setuptools
six
types-dataclasses
typing_extensions
sympy
tqdm
matplotlib
seaborn
flasgger~=0.9.5
